+++ src/app/features/links/link-view/link-view.component.ts
@@ -73,14 +73,14 @@ export class LinkViewComponent implements OnInit, OnDestroy {
     const SHARED_URL = `${baseUrl}${routerUrl}`;
 
     switch ( social ) {
-      case 'twitter':
-        const title = `${this.link.name} | Rincón de Soto`;
-        window.open('http://twitter.com/share?url='+encodeURIComponent(SHARED_URL)+'&text='+encodeURIComponent(title), '', 'left=0,top=0,width=550,height=450,personalbar=0,toolbar=0,scrollbars=0,resizable=0');
-        break;
+      // case 'twitter':
+      //   const title = `${this.link.name} | Rincón de Soto`;
+      //   window.open('http://twitter.com/share?url='+encodeURIComponent(SHARED_URL)+'&text='+encodeURIComponent(title), '', 'left=0,top=0,width=550,height=450,personalbar=0,toolbar=0,scrollbars=0,resizable=0');
+      //   break;
 
-      case 'facebook':
-        window.open('http://facebook.com/sharer/sharer.php?u='+encodeURIComponent(SHARED_URL), '', 'left=0,top=0,width=650,height=420,personalbar=0,toolbar=0,scrollbars=0,resizable=0');
-        break;
+      // case 'facebook':
+      //   window.open('http://facebook.com/sharer/sharer.php?u='+encodeURIComponent(SHARED_URL), '', 'left=0,top=0,width=650,height=420,personalbar=0,toolbar=0,scrollbars=0,resizable=0');
+      //   break;
 
       case 'whatsapp':
         window.open(`whatsapp://send?text=${SHARED_URL}`);